Revert "update expectation for collstats"
This reverts commit a31f3e44e866ec44b02e0fae5e2283aeb4419585.

+++ internal/handlers/sqlite/msg_collstats.go
@@ -110,8 +110,8 @@ func (h *Handler) MsgCollStats(ctx context.Context, msg *wire.OpMsg) (*wire.OpMs
 
 	if capped {
 		pairs = append(pairs,
-			"max", int32(stats.CappedDocuments),
-			"maxSize", int32(stats.CappedSize/scale),
+			"max", stats.CappedDocuments,
+			"maxSize", stats.CappedSize/scale,
 		)
 	}
